refactor(checker): split analysis-state value types out of context/mod.rs under the 2000-LOC cap
context/mod.rs reached 2030 physical lines on main (over the tsz-checker/src
hard 2000 cap), introduced by a native-queue merge that never ran the local
arch guard. Extract the in-progress analysis-state value types
(PendingImplicitAnyVar, PendingCircularReturnSites, InferredReturnTypeKey,
PartialObjectLiteralInitializer, ObjectLiteralTracking) plus their own inherent
impls to a new child module context/analysis_state_types.rs, re-exported from
the module root (pub use analysis_state_types::*) so every crate::context::*
call site is unchanged. Pure code motion: types stay pub with pub fields, the
child reads super::PendingImplicitAnyKind, and mod.rs drops its now-unused
PropertyInfo import. context/mod.rs 2030 -> 1937.
